+++ code/src/sixsq/nuvla/server/resources/customer/utils.clj
@@ -185,9 +185,11 @@
   [customer-id {:keys [plan-id plan-item-ids] :as body} trial?]
   (let [catalogue (crud/retrieve-by-id-as-admin pricing/resource-id)]
     (throw-plan-invalid body catalogue)
-    (-> {"customer" customer-id
-         "items"    (map (fn [plan-id] {"price" plan-id})
-                         (cons plan-id plan-item-ids))}
+    (-> {"customer"          customer-id
+         "items"             (map (fn [plan-id] {"price" plan-id})
+                                  (cons plan-id plan-item-ids))
+         "collection_method" "send_invoice"
+         "days_until_due"    14}
         (cond-> trial? (assoc "trial_period_days" 14))
         stripe/create-subscription
         s-subscription->map)))