+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
   - New cimi common attributes to store user-id
     (created-by, updated-by)
+  - New optional nuvlabox-status resource for GPIO telemetry
 
 ### Changed