+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changed
 
+  - New optional nuvlabox-status attribute for GPIO telemetry
+  - New optional nuvlabox-engine-version attribute in nuvlabox-status
   - Customer - Susbscription send invoices by default
   - Switch group - allow switch to group/nuvla-admin
     user super has to do a switch to do administative tasks
@@ -18,7 +20,6 @@
 
   - New cimi common attributes to store user-id
     (created-by, updated-by)
-  - New optional nuvlabox-status resource for GPIO telemetry
 
 ### Changed